refactor(phase4d): sample 画像生成系を library/sampling.py に分離
train_util.py から以下を library/sampling.py に移動 (498 行):
- noise scheduler 定数 (SCHEDULER_LINEAR_START / _END / _TIMESTEPS / SCHEDLER_SCHEDULE)
- get_my_scheduler (--sample_sampler 名 → diffusers Scheduler ファクトリ)
- line_to_prompt_dict / load_prompts (.txt / .toml / .json プロンプト解析)
- sample_images / sample_images_common / sample_image_inference
  (multi-GPU 分配・RNG 退避・inpainting 入力・W&B 画像 upload)

train_util.py は re-export shim を維持。当リポジトリ内 16 トップレベル
+ 5 library/*_train_util.py = 21 ファイルの参照を sampling.* (alias) に切替。

循環依存対策: sampling.py → sdxl_lpw_stable_diffusion → sdxl_train_util →
train_util → sampling のループになるため、SdxlStableDiffusionLongPromptWeightingPipeline
は sample_image_inference 内で遅延 import (isinstance チェックでのみ使用)。

train_util.py: 975 → 555 行 (420 行削減)。
