+++ library/sdxl_train_util.py
@@ -11,7 +11,7 @@ init_ipex()
 from accelerate import init_empty_weights
 from tqdm import tqdm
 from transformers import CLIPTokenizer
-from library import model_util, sdxl_model_util, checkpoint_io, train_util, sdxl_original_unet
+from library import model_util, sdxl_model_util, checkpoint_io, sampling, train_util, sdxl_original_unet
 from .utils import setup_logging
 
 setup_logging()
@@ -390,4 +390,4 @@ def verify_sdxl_training_args(args: argparse.Namespace, support_text_encoder_cac
 def sample_images(*args, **kwargs):
     from library.sdxl_lpw_stable_diffusion import SdxlStableDiffusionLongPromptWeightingPipeline
 
-    return train_util.sample_images_common(SdxlStableDiffusionLongPromptWeightingPipeline, *args, **kwargs)
+    return sampling.sample_images_common(SdxlStableDiffusionLongPromptWeightingPipeline, *args, **kwargs)
