+++ library/train_util.py
@@ -475,441 +475,21 @@ from library.loss import (  # noqa: F401
 
 
 
-# scheduler:
-SCHEDULER_LINEAR_START = 0.00085
-SCHEDULER_LINEAR_END = 0.0120
-SCHEDULER_TIMESTEPS = 1000
-SCHEDLER_SCHEDULE = "scaled_linear"
-
-
-def get_my_scheduler(
-    *,
-    sample_sampler: str,
-    v_parameterization: bool,
-):
-    sched_init_args = {}
-    if sample_sampler == "ddim":
-        scheduler_cls = DDIMScheduler
-    elif sample_sampler == "ddpm":  # ddpmはおかしくなるのでoptionから外してある
-        scheduler_cls = DDPMScheduler
-    elif sample_sampler == "pndm":
-        scheduler_cls = PNDMScheduler
-    elif sample_sampler == "lms" or sample_sampler == "k_lms":
-        scheduler_cls = LMSDiscreteScheduler
-    elif sample_sampler == "euler" or sample_sampler == "k_euler":
-        scheduler_cls = EulerDiscreteScheduler
-    elif sample_sampler == "euler_a" or sample_sampler == "k_euler_a":
-        scheduler_cls = EulerAncestralDiscreteScheduler
-    elif sample_sampler == "dpmsolver" or sample_sampler == "dpmsolver++":
-        scheduler_cls = DPMSolverMultistepScheduler
-        sched_init_args["algorithm_type"] = sample_sampler
-    elif sample_sampler == "dpmsingle":
-        scheduler_cls = DPMSolverSinglestepScheduler
-    elif sample_sampler == "heun":
-        scheduler_cls = HeunDiscreteScheduler
-    elif sample_sampler == "dpm_2" or sample_sampler == "k_dpm_2":
-        scheduler_cls = KDPM2DiscreteScheduler
-    elif sample_sampler == "dpm_2_a" or sample_sampler == "k_dpm_2_a":
-        scheduler_cls = KDPM2AncestralDiscreteScheduler
-    else:
-        scheduler_cls = DDIMScheduler
-
-    if v_parameterization:
-        sched_init_args["prediction_type"] = "v_prediction"
-
-    scheduler = scheduler_cls(
-        num_train_timesteps=SCHEDULER_TIMESTEPS,
-        beta_start=SCHEDULER_LINEAR_START,
-        beta_end=SCHEDULER_LINEAR_END,
-        beta_schedule=SCHEDLER_SCHEDULE,
-        steps_offset=1,
-        **sched_init_args,
-    )
-
-    # clip_sample=Trueにする
-    if hasattr(scheduler.config, "clip_sample") and scheduler.config.clip_sample is False:
-        # logger.info("set clip_sample to True")
-        scheduler.config.clip_sample = True
-
-    return scheduler
-
-
-def sample_images(*args, **kwargs):
-    return sample_images_common(StableDiffusionLongPromptWeightingPipeline, *args, **kwargs)
-
-
-def line_to_prompt_dict(line: str) -> dict:
-    # subset of gen_img_diffusers
-    prompt_args = line.split(" --")
-    prompt_dict = {}
-    prompt_dict["prompt"] = prompt_args[0]
-
-    for parg in prompt_args:
-        try:
-            m = re.match(r"w (\d+)", parg, re.IGNORECASE)
-            if m:
-                prompt_dict["width"] = int(m.group(1))
-                continue
-
-            m = re.match(r"h (\d+)", parg, re.IGNORECASE)
-            if m:
-                prompt_dict["height"] = int(m.group(1))
-                continue
-
-            m = re.match(r"d (\d+)", parg, re.IGNORECASE)
-            if m:
-                prompt_dict["seed"] = int(m.group(1))
-                continue
-
-            m = re.match(r"s (\d+)", parg, re.IGNORECASE)
-            if m:  # steps
-                prompt_dict["sample_steps"] = max(1, min(1000, int(m.group(1))))
-                continue
-
-            m = re.match(r"l ([\d\.]+)", parg, re.IGNORECASE)
-            if m:  # scale
-                prompt_dict["scale"] = float(m.group(1))
-                continue
-
-            m = re.match(r"g ([\d\.]+)", parg, re.IGNORECASE)
-            if m:  # guidance scale
-                prompt_dict["guidance_scale"] = float(m.group(1))
-                continue
-
-            m = re.match(r"n (.+)", parg, re.IGNORECASE)
-            if m:  # negative prompt
-                prompt_dict["negative_prompt"] = m.group(1)
-                continue
-
-            m = re.match(r"ss (.+)", parg, re.IGNORECASE)
-            if m:
-                prompt_dict["sample_sampler"] = m.group(1)
-                continue
-
-            m = re.match(r"cn (.+)", parg, re.IGNORECASE)
-            if m:
-                prompt_dict["controlnet_image"] = m.group(1)
-                continue
-
-            m = re.match(r"i (.+)", parg, re.IGNORECASE)
-            if m:
-                prompt_dict["image"] = m.group(1).strip()
-                continue
-
-            m = re.match(r"ctr (.+)", parg, re.IGNORECASE)
-            if m:
-                prompt_dict["cfg_trunc_ratio"] = float(m.group(1))
-                continue
-
-            m = re.match(r"rcfg (.+)", parg, re.IGNORECASE)
-            if m:
-                prompt_dict["renorm_cfg"] = float(m.group(1))
-                continue
-
-            m = re.match(r"fs (.+)", parg, re.IGNORECASE)
-            if m:
-                prompt_dict["flow_shift"] = m.group(1)
-                continue
-
-            m = re.match(r"am ([\d\.\-,]+)", parg, re.IGNORECASE)
-            if m:  # additional network multiplier(s) — comma-separated list, same as gen_img.py
-                prompt_dict["additional_network_multiplier"] = [float(v) for v in m.group(1).split(",")]
-                continue
-
-        except ValueError as ex:
-            logger.error(f"Exception in parsing / 解析エラー: {parg}")
-            logger.error(ex)
-
-    return prompt_dict
-
-
-def load_prompts(prompt_file: str) -> List[Dict]:
-    # read prompts
-    if prompt_file.endswith(".txt"):
-        with open(prompt_file, "r", encoding="utf-8") as f:
-            lines = f.readlines()
-        prompts = [line.strip() for line in lines if len(line.strip()) > 0 and line[0] != "#"]
-    elif prompt_file.endswith(".toml"):
-        with open(prompt_file, "r", encoding="utf-8") as f:
-            data = toml.load(f)
-        prompts = [dict(**data["prompt"], **subset) for subset in data["prompt"]["subset"]]
-    elif prompt_file.endswith(".json"):
-        with open(prompt_file, "r", encoding="utf-8") as f:
-            prompts = json.load(f)
-
-    # preprocess prompts
-    for i in range(len(prompts)):
-        prompt_dict = prompts[i]
-        if isinstance(prompt_dict, str):
-            from library.train_util import line_to_prompt_dict
-
-            prompt_dict = line_to_prompt_dict(prompt_dict)
-            prompts[i] = prompt_dict
-        assert isinstance(prompt_dict, dict)
-
-        # Adds an enumerator to the dict based on prompt position. Used later to name image files. Also cleanup of extra data in original prompt dict.
-        prompt_dict["enum"] = i
-        prompt_dict.pop("subset", None)
-
-    return prompts
-
-
-def sample_images_common(
-    pipe_class,
-    accelerator: Accelerator,
-    args: argparse.Namespace,
-    epoch: int,
-    steps: int,
-    device,
-    vae,
-    tokenizer,
-    text_encoder,
-    unet_wrapped,
-    prompt_replacement=None,
-    controlnet=None,
-):
-    """
-    StableDiffusionLongPromptWeightingPipelineの改造版を使うようにしたので、clip skipおよびプロンプトの重みづけに対応した
-    TODO Use strategies here
-    """
-
-    if steps == 0:
-        if not args.sample_at_first:
-            return
-    else:
-        if args.sample_every_n_steps is None and args.sample_every_n_epochs is None:
-            return
-        if args.sample_every_n_epochs is not None:
-            # sample_every_n_steps は無視する
-            if epoch is None or epoch % args.sample_every_n_epochs != 0:
-                return
-        else:
-            if steps % args.sample_every_n_steps != 0 or epoch is not None:  # steps is not divisible or end of epoch
-                return
-
-    logger.info("")
-    logger.info(f"generating sample images at step / サンプル画像生成 ステップ: {steps}")
-    if not os.path.isfile(args.sample_prompts):
-        logger.error(f"No prompt file / プロンプトファイルがありません: {args.sample_prompts}")
-        return
-
-    distributed_state = PartialState()  # for multi gpu distributed inference. this is a singleton, so it's safe to use it here
-
-    org_vae_device = vae.device  # CPUにいるはず
-    vae.to(distributed_state.device)  # distributed_state.device is same as accelerator.device
-
-    # unwrap unet and text_encoder(s)
-    unet = accelerator.unwrap_model(unet_wrapped)
-    if isinstance(text_encoder, (list, tuple)):
-        text_encoder = [accelerator.unwrap_model(te) for te in text_encoder]
-    else:
-        text_encoder = accelerator.unwrap_model(text_encoder)
-
-    # read prompts
-    if args.sample_prompts.endswith(".txt"):
-        with open(args.sample_prompts, "r", encoding="utf-8") as f:
-            lines = f.readlines()
-        prompts = [line.strip() for line in lines if len(line.strip()) > 0 and line[0] != "#"]
-    elif args.sample_prompts.endswith(".toml"):
-        with open(args.sample_prompts, "r", encoding="utf-8") as f:
-            data = toml.load(f)
-        prompts = [dict(**data["prompt"], **subset) for subset in data["prompt"]["subset"]]
-    elif args.sample_prompts.endswith(".json"):
-        with open(args.sample_prompts, "r", encoding="utf-8") as f:
-            prompts = json.load(f)
-
-    default_scheduler = get_my_scheduler(sample_sampler=args.sample_sampler, v_parameterization=args.v_parameterization)
-
-    pipeline = pipe_class(
-        text_encoder=text_encoder,
-        vae=vae,
-        unet=unet,
-        tokenizer=tokenizer,
-        scheduler=default_scheduler,
-        safety_checker=None,
-        feature_extractor=None,
-        requires_safety_checker=False,
-        clip_skip=args.clip_skip,
-    )
-    pipeline.to(distributed_state.device)
-    save_dir = args.output_dir + "/sample"
-    os.makedirs(save_dir, exist_ok=True)
-
-    # preprocess prompts
-    for i in range(len(prompts)):
-        prompt_dict = prompts[i]
-        if isinstance(prompt_dict, str):
-            prompt_dict = line_to_prompt_dict(prompt_dict)
-            prompts[i] = prompt_dict
-        assert isinstance(prompt_dict, dict)
-
-        # Adds an enumerator to the dict based on prompt position. Used later to name image files. Also cleanup of extra data in original prompt dict.
-        prompt_dict["enum"] = i
-        prompt_dict.pop("subset", None)
-
-    # save random state to restore later
-    rng_state = torch.get_rng_state()
-    cuda_rng_state = None
-    try:
-        cuda_rng_state = torch.cuda.get_rng_state() if torch.cuda.is_available() else None
-    except Exception:
-        pass
-
-    if distributed_state.num_processes <= 1:
-        # If only one device is available, just use the original prompt list. We don't need to care about the distribution of prompts.
-        with torch.no_grad():
-            for prompt_dict in prompts:
-                sample_image_inference(
-                    accelerator, args, pipeline, save_dir, prompt_dict, epoch, steps, prompt_replacement, controlnet=controlnet
-                )
-    else:
-        # Creating list with N elements, where each element is a list of prompt_dicts, and N is the number of processes available (number of devices available)
-        # prompt_dicts are assigned to lists based on order of processes, to attempt to time the image creation time to match enum order. Probably only works when steps and sampler are identical.
-        per_process_prompts = []  # list of lists
-        for i in range(distributed_state.num_processes):
-            per_process_prompts.append(prompts[i :: distributed_state.num_processes])
-
-        with torch.no_grad():
-            with distributed_state.split_between_processes(per_process_prompts) as prompt_dict_lists:
-                for prompt_dict in prompt_dict_lists[0]:
-                    sample_image_inference(
-                        accelerator, args, pipeline, save_dir, prompt_dict, epoch, steps, prompt_replacement, controlnet=controlnet
-                    )
-
-    # clear pipeline and cache to reduce vram usage
-    del pipeline
-
-    torch.set_rng_state(rng_state)
-    if torch.cuda.is_available() and cuda_rng_state is not None:
-        torch.cuda.set_rng_state(cuda_rng_state)
-    vae.to(org_vae_device)
-
-    clean_memory_on_device(accelerator.device)
-
-
-def sample_image_inference(
-    accelerator: Accelerator,
-    args: argparse.Namespace,
-    pipeline: Union[StableDiffusionLongPromptWeightingPipeline, SdxlStableDiffusionLongPromptWeightingPipeline],
-    save_dir,
-    prompt_dict,
-    epoch,
-    steps,
-    prompt_replacement,
-    controlnet=None,
-):
-    assert isinstance(prompt_dict, dict)
-    negative_prompt = prompt_dict.get("negative_prompt")
-    sample_steps = prompt_dict.get("sample_steps", 30)
-    width = prompt_dict.get("width", 512)
-    height = prompt_dict.get("height", 512)
-    scale = prompt_dict.get("scale", 7.5)
-    seed = prompt_dict.get("seed")
-    controlnet_image = prompt_dict.get("controlnet_image")
-    prompt: str = prompt_dict.get("prompt", "")
-    sampler_name: str = prompt_dict.get("sample_sampler", args.sample_sampler)
-
-    if prompt_replacement is not None:
-        prompt = prompt.replace(prompt_replacement[0], prompt_replacement[1])
-        if negative_prompt is not None:
-            negative_prompt = negative_prompt.replace(prompt_replacement[0], prompt_replacement[1])
-
-    if seed is not None:
-        torch.manual_seed(seed)
-        if torch.cuda.is_available():
-            torch.cuda.manual_seed(seed)
-    else:
-        # True random sample image generation
-        torch.seed()
-        if torch.cuda.is_available():
-            torch.cuda.seed()
-
-    scheduler = get_my_scheduler(
-        sample_sampler=sampler_name,
-        v_parameterization=args.v_parameterization,
-    )
-    pipeline.scheduler = scheduler
-
-    if controlnet_image is not None:
-        controlnet_image = Image.open(controlnet_image).convert("RGB")
-        controlnet_image = controlnet_image.resize((width, height), Image.LANCZOS)
-
-    # Round down so the latent shape is divisible by the UNet's largest stride:
-    # SDXL has 2 downsamples (latent /4 → image /32); SD1.5/2.x has 3 (latent /8 → image /64).
-    divisor = 32 if isinstance(pipeline, SdxlStableDiffusionLongPromptWeightingPipeline) else 64
-    height = max(divisor, height - height % divisor)
-    width = max(divisor, width - width % divisor)
-    logger.info(f"prompt: {prompt}")
-    logger.info(f"negative_prompt: {negative_prompt}")
-    logger.info(f"height: {height}")
-    logger.info(f"width: {width}")
-    logger.info(f"sample_steps: {sample_steps}")
-    logger.info(f"scale: {scale}")
-    logger.info(f"sample_sampler: {sampler_name}")
-    if seed is not None:
-        logger.info(f"seed: {seed}")
-
-    # Prepare inpainting source image and mask when training an inpainting model.
-    # The prompt line should include "--i /path/to/image.jpg".
-    # The mask is generated reproducibly from the prompt seed (or randomly if no seed).
-    inpaint_image = None
-    inpaint_mask = None
-    if getattr(args, "train_inpainting", False):
-        image_path = prompt_dict.get("image")
-        if image_path:
-            from library.mask_generator import wobbly_ellipse_mask as _gen_mask
-
-            if not os.path.exists(image_path):
-                logger.warning(f"inpaint image not found, skipping sample: {image_path}")
-                return
-            inpaint_image = Image.open(image_path).convert("RGB").resize((width, height), Image.LANCZOS)
-            inpaint_mask = _gen_mask(width, height, seed=seed)
-            logger.info(f"inpaint image: {image_path}")
-        else:
-            logger.warning(
-                "train_inpainting is set but no source image specified in the prompt line; "
-                "skipping sample. Add '--i /path/to/image.jpg' to the prompt to enable inpainting sampling."
-            )
-            return
-
-    with accelerator.autocast(), torch.no_grad():
-        latents = pipeline(
-            prompt=prompt,
-            height=height,
-            width=width,
-            num_inference_steps=sample_steps,
-            guidance_scale=scale,
-            negative_prompt=negative_prompt,
-            controlnet=controlnet,
-            controlnet_image=controlnet_image,
-            inpaint_image=inpaint_image,
-            inpaint_mask=inpaint_mask,
-        )
-
-    if torch.cuda.is_available():
-        with torch.cuda.device(torch.cuda.current_device()):
-            torch.cuda.empty_cache()
-
-    image = pipeline.latents_to_image(latents)[0]
-
-    # adding accelerator.wait_for_everyone() here should sync up and ensure that sample images are saved in the same order as the original prompt list
-    # but adding 'enum' to the filename should be enough
-
-    ts_str = time.strftime("%Y%m%d%H%M%S", time.localtime())
-    num_suffix = f"e{epoch:06d}" if epoch is not None else f"{steps:06d}"
-    seed_suffix = "" if seed is None else f"_{seed}"
-    i: int = prompt_dict["enum"]
-    img_filename = f"{'' if args.output_name is None else args.output_name + '_'}{num_suffix}_{i:02d}_{ts_str}{seed_suffix}.png"
-    image.save(os.path.join(save_dir, img_filename))
-
-    # send images to wandb if enabled
-    if "wandb" in [tracker.name for tracker in accelerator.trackers]:
-        wandb_tracker = accelerator.get_tracker("wandb")
-
-        import wandb
-
-        # not to commit images to avoid inconsistency between training and logging steps
-        wandb_tracker.log({f"sample_{i}": wandb.Image(image, caption=prompt)}, commit=False)  # positive prompt as a caption
+# Sampling helpers (default scheduler, prompt parsing, sample generation)
+# have moved to library.sampling; re-exported here for backward compatibility.
+# New code should import from library.sampling directly.
+from library.sampling import (  # noqa: F401, E402
+    SCHEDULER_LINEAR_START,
+    SCHEDULER_LINEAR_END,
+    SCHEDULER_TIMESTEPS,
+    SCHEDLER_SCHEDULE,
+    get_my_scheduler,
+    sample_images,
+    line_to_prompt_dict,
+    load_prompts,
+    sample_images_common,
+    sample_image_inference,
+)
 
 
 # Logging / tracker helpers have moved to library.logging_util;
